refactor(analytics): split analytics.ts into focused modules
Extracts the 2245-line analytics.ts into four focused modules:

- analytics-types.ts: shared types, interfaces, window augmentation,
  and capability constants (DeploymentType, InstallCopySource, UtmParams,
  ProviderSummary, SendOptions)

- analytics-session.ts: bot/headless detection, client & session identity
  (CID/SID/SC/LAST keys, getClientId, getSession), anonymous user ID
  hashing, engagement time tracking state and functions, UTM param loading

- analytics-core.ts: dual-path send pipeline (gtag.js primary, custom
  proxy fallback), Umami parallel tracking, event queuing while waiting
  for gtag.js load verdict, initAnalytics, updateAnalyticsIds,
  setAnalyticsUserId/UserProperties, setAnalyticsOptOut, page view,
  error tracking (emitError, markErrorReported, startGlobalErrorTracking,
  chunk-reload recovery)

- analytics-events.ts: all domain-specific emit* functions organized
  by product area (dashboard, search, missions, auth, feedback, NPS,
  marketplace, cluster, agent, deploy, compliance, widgets, nudges,
  tour, ACMM, landing pages, etc.)

analytics.ts becomes a zero-logic barrel re-exporting every symbol from
the four modules. All existing imports from 'lib/analytics' continue to
work without any changes to callers.

Fixes #8624
